+++ aeneas/cew/cew_setup.py
@@ -47,7 +47,7 @@ CMODULE = Extension(
 
 setup(
     name="cew",
-    version="1.6.0",
+    version="1.6.1",
     description="Python C Extension for synthesizing text with eSpeak.",
     ext_modules=[CMODULE]
 )